+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/), adapted for int
 
 ## [Unreleased]
 
+## [397] - 2026-03-12
+
+- [TASK-579] Fix tusk commit pathspec error when paths are relative to a subdirectory
+
 ## [396] - 2026-03-12
 
 - [TASK-578] Fix tusk commit false-positive lint-staged failure when commit succeeds